+++ PythonRouge.game/Colours.cs
@@ -24,7 +24,7 @@ namespace PythonRouge.game
     /// <summary>
     /// This is a class designed to hold custom colours defined by me.
     /// </summary>
-    static class Colours
+    public static class Colours
     {
         public static RLColor floor_lit = new RLColor(68, 68, 63);
         public static RLColor wall_lit = new RLColor(53, 71, 36);
