+++ PythonRouge/Program.cs
@@ -10,9 +10,8 @@
 // You should have received a copy of the GNU General Public License along with this program. If not, see
 // http://www.gnu.org/licenses/.
 
-using System;
 using PythonRouge.game;
-using PythonRouge.network;
+using PythonRouge.Network;
 using RLNET;
 
 namespace PythonRouge
@@ -58,30 +57,6 @@ namespace PythonRouge
             _rootConsole.Run();
         }
 
-        //Main menu this just prints the options available for the main menu.
-        private static void mainMenu()
-        {
-            _rootConsole.Print(36, 1, "Press e to exit", RLColor.White);
-            _rootConsole.Print(4, 3, "Game", RLColor.Cyan);
-            _rootConsole.Print(4, 4, "1) Play Game", RLColor.White);
-            _rootConsole.Print(4, 5, "2) Multiplayer", RLColor.White);
-        }
-
-        //The same as the above only for the multiplayer menu.
-        private static void multiMenu()
-        {
-            _rootConsole.Print(36, 1, "Press e to exit", RLColor.White);
-            _rootConsole.Print(4, 3, "Multiplayer", RLColor.Cyan);
-            _rootConsole.Print(4, 4, "1) Host Game", RLColor.White);
-            _rootConsole.Print(4, 5, "2) Join Game", RLColor.White);
-            multiRendered = true;
-        }
-
-        private static void enterDets()
-        {
-            _rootConsole.Print(4, 3, "Enter Name:", RLColor.White);
-            _rootConsole.Print(15, 3, name, RLColor.White);
-        }
 
         /// <summary>
         ///     This is the event handler for the Root Console update event.
@@ -93,314 +68,54 @@ namespace PythonRouge
         /// <param name="e"></param>
         private static void rootConsole_Update(object sender, UpdateEventArgs e)
         {
-            if (!wantName)
+            
+           if(Menu.currMenu == "main")
             {
-                if (_spEngine != null)
+                switch(Menu.mainUpdate(_rootConsole))
                 {
-                    var keyPress = _rootConsole.Keyboard.GetKeyPress();
-                    if (keyPress != null)
-                        _spEngine.handleKey(keyPress);
+                    case 1:
+                        _spEngine = new SPEngine(_rootConsole);
+                        break;
+                    case 2:
+                        Menu.currMenu = "multi";
+                        break;
                 }
-                if (_mpEngine != null)
+            }
+           if(Menu.currMenu == "multi")
+            {
+                switch(Menu.multiUpdate(_rootConsole))
                 {
-                    var keyPress = _rootConsole.Keyboard.GetKeyPress();
-                    if (keyPress != null)
-                        _mpEngine.HandleKey(keyPress);
-                    _mpEngine.Update();
+                    case 1:
+                        break;
+                    case 2:
+                        Menu.currMenu = "enterDets";
+                        Menu.name = "";
+                        break;
                 }
-                var keyPrss = _rootConsole.Keyboard.GetKeyPress();
-                if (keyPrss != null)
+            }
+           if(Menu.currMenu == "enterDets")
+            {
+                switch(Menu.enterDetsUpdate(_rootConsole))
                 {
-                    if (keyPrss.Key == RLKey.Number1)
-                    {
-                        if (main)
-                            _spEngine = new SPEngine(_rootConsole);
+                    case 1:
+                        _mpEngine = new MpEngine(_rootConsole, Menu.name);
+                        Menu.currMenu = "game";
+                        break;
+                    case 0:
+                        break;
+                }
+            }
+            if (Menu.currMenu == "game")
+            {
+                var keypress = _rootConsole.Keyboard.GetKeyPress();
+                if (keypress != null)
+                {
+                    _spEngine?.handleKey(keypress);
+                    _mpEngine?.HandleKey(keypress);
+                }
+                _mpEngine?.Update();
+            }
 
-                    }
-                    if (keyPrss.Key == RLKey.Number2)
-                        if (main)
-                        {
-                            main = false;
-                            Multi = true;
-                            wantName = true;
-                        }
-                    if (Multi && multiRendered && detsEntered)
-                        _mpEngine = new MpEngine(_rootConsole, name);
-                    if (keyPrss.Key == RLKey.E)
-                    {
-                        if (Multi)
-                        {
-                            main = true;
-                            Multi = false;
-                        }
-                        if (main)
-                            _rootConsole.Close();
-                    }
-                }
-            }
-            if (!detsEntered && wantName)
-            {
-                var key = _rootConsole.Keyboard.GetKeyPress();
-                if (key != null)
-                {
-                    switch (key.Key)
-                    {
-                        case RLKey.Enter:
-                            detsEntered = true;
-                            wantName = false;
-                            break;
-                        case RLKey.Escape:
-                            break;
-                        case RLKey.Space:
-                            name += " ";
-                            break;
-                        case RLKey.Back:
-                            if(name.Length != 0)
-                            name = name.Remove(name.Length - 1);
-                            break;
-                        case RLKey.A:
-                            if (key.Shift || key.CapsLock)
-                                name += "A";
-                            else
-                            {
-                                name += "a";
-                            }
-                            break;
-                        case RLKey.B:
-                            if (key.Shift || key.CapsLock)
-                                name += "B";
-                            else
-                            {
-                                name += "b";
-                            }
-                            break;
-                        case RLKey.C:
-                            if (key.Shift || key.CapsLock)
-                                name += "C";
-                            else
-                            {
-                                name += "c";
-                            }
-                            break;
-                        case RLKey.D:
-                            if (key.Shift || key.CapsLock)
-                                name += "D";
-                            else
-                            {
-                                name += "d";
-                            }
-                            break;
-                        case RLKey.E:
-                            if (key.Shift || key.CapsLock)
-                                name += "E";
-                            else
-                            {
-                                name += "e";
-                            }
-                            break;
-                        case RLKey.F:
-                            if (key.Shift || key.CapsLock)
-                                name += "F";
-                            else
-                            {
-                                name += "f";
-                            }
-                            break;
-                        case RLKey.G:
-                            if (key.Shift || key.CapsLock)
-                                name += "G";
-                            else
-                            {
-                                name += "g";
-                            }
-                            break;
-                        case RLKey.H:
-                            if (key.Shift || key.CapsLock)
-                                name += "H";
-                            else
-                            {
-                                name += "h";
-                            }
-                            break;
-                        case RLKey.I:
-                            if (key.Shift || key.CapsLock)
-                                name += "I";
-                            else
-                            {
-                                name += "i";
-                            }
-                            break;
-                        case RLKey.J:
-                            if (key.Shift || key.CapsLock)
-                                name += "J";
-                            else
-                            {
-                                name += "j";
-                            }
-                            break;
-                        case RLKey.K:
-                            if (key.Shift || key.CapsLock)
-                                name += "K";
-                            else
-                            {
-                                name += "k";
-                            }
-                            break;
-                        case RLKey.L:
-                            if (key.Shift || key.CapsLock)
-                                name += "L";
-                            else
-                            {
-                                name += "l";
-                            }
-                            break;
-                        case RLKey.M:
-                            if (key.Shift || key.CapsLock)
-                                name += "M";
-                            else
-                            {
-                                name += "m";
-                            }
-                            break;
-                        case RLKey.N:
-                            if (key.Shift || key.CapsLock)
-                                name += "N";
-                            else
-                            {
-                                name += "n";
-                            }
-                            break;
-                        case RLKey.O:
-                            if (key.Shift || key.CapsLock)
-                                name += "O";
-                            else
-                            {
-                                name += "o";
-                            }
-                            break;
-                        case RLKey.P:
-                            if (key.Shift || key.CapsLock)
-                                name += "P";
-                            else
-                            {
-                                name += "p";
-                            }
-                            break;
-                        case RLKey.Q:
-                            if (key.Shift || key.CapsLock)
-                                name += "Q";
-                            else
-                            {
-                                name += "q";
-                            }
-                            break;
-                        case RLKey.R:
-                            if (key.Shift || key.CapsLock)
-                                name += "R";
-                            else
-                            {
-                                name += "r";
-                            }
-                            break;
-                        case RLKey.S:
-                            if (key.Shift || key.CapsLock)
-                                name += "S";
-                            else
-                            {
-                                name += "s";
-                            }
-                            break;
-                        case RLKey.T:
-                            if (key.Shift || key.CapsLock)
-                                name += "T";
-                            else
-                            {
-                                name += "t";
-                            }
-                            break;
-                        case RLKey.U:
-                            if (key.Shift || key.CapsLock)
-                                name += "U";
-                            else
-                            {
-                                name += "u";
-                            }
-                            break;
-                        case RLKey.V:
-                            if (key.Shift || key.CapsLock)
-                                name += "V";
-                            else
-                            {
-                                name += "v";
-                            }
-                            break;
-                        case RLKey.W:
-                            if (key.Shift || key.CapsLock)
-                                name += "W";
-                            else
-                            {
-                                name += "w";
-                            }
-                            break;
-                        case RLKey.X:
-                            if (key.Shift || key.CapsLock)
-                                name += "X";
-                            else
-                            {
-                                name += "x";
-                            }
-                            break;
-                        case RLKey.Y:
-                            if (key.Shift || key.CapsLock)
-                                name += "Y";
-                            else
-                            {
-                                name += "y";
-                            }
-                            break;
-                        case RLKey.Z:
-                            if (key.Shift || key.CapsLock)
-                                name += "Z";
-                            else
-                            {
-                                name += "z";
-                            }
-                            break;
-                        case RLKey.Number0:
-                            name += "0";
-                            break;
-                        case RLKey.Number1:
-                            name += "1";
-                            break;
-                        case RLKey.Number2:
-                            name += "2";
-                            break;
-                        case RLKey.Number3:
-                            name += "3";
-                            break;
-                        case RLKey.Number4:
-                            name += "4";
-                            break;
-                        case RLKey.Number5:
-                            name += "5";
-                            break;
-                        case RLKey.Number6:
-                            name += "6";
-                            break;
-                        case RLKey.Number7:
-                            name += "7";
-                            break;
-                        case RLKey.Number8:
-                            name += "8";
-                            break;
-                        case RLKey.Number9:
-                            name += "9";
-                            break;
-                        default:
-                            break;
-                    }
-                }
-            }
         }
 
         /// <summary>
@@ -412,19 +127,26 @@ namespace PythonRouge
         private static void rootConsole_Render(object sender, UpdateEventArgs e)
         {
             _rootConsole.Clear();
-            _spEngine?.render();
-            _mpEngine?.Render();
-            if (_spEngine == null && _mpEngine == null)
+            switch (Menu.currMenu)
             {
-                if (main)
-                    mainMenu();
-                if (Multi && !wantName)
-                    multiMenu();
-                if (!detsEntered && wantName)
-                {
-                    enterDets();
-                }
+                case "main":
+                    Menu.mainRender(_rootConsole);
+                    break;
+                case "multi":
+                    Menu.multiRender(_rootConsole);
+                    break;
+                case "enterDets":
+                    Menu.enterDetsRender(_rootConsole);
+                    break;
+                case "game":
+                    _spEngine?.render();
+                    _mpEngine?.Render();
+                    break;
+
+
             }
+            
+            
             _rootConsole.Draw();
         }
     }
